Moved wiring graph to event page, fixed wiring counter bug
