+++ statechanges/templates/statechanges/eventgraph.html
@@ -1,0 +1,59 @@
+<script src="https://cdn.jsdelivr.net/npm/chart.js@2.9.3/dist/Chart.bundle.min.js"></script>
+<script>
+    var ctx = document.getElementById('ChartWiringsDay').getContext('2d');
+    Chart.defaults.global.defaultFontColor = '#ffffff';
+    Chart.defaults.global.defaultFontFamily = '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,"Noto Sans",sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol","Noto Color Emoji"';
+    Chart.defaults.global.defaultFontSize = 16;
+    var chart = new Chart(ctx, {
+        type: 'line',
+        data: {
+            labels: [
+                {% for info in wiringgraphinfo %}
+                '{{info.periodname}}'{% if not forloop.last %}, {% endif %}
+    {% endfor %}
+            ],
+    datasets: [{
+        label: 'New events published in the last 30 days)',
+        backgroundColor: '#000000',
+        borderColor: '#01C696',
+        data: [
+            {% for info in wiringgraphinfo %}
+                    '{{info.value}}'{% if not forloop.last %}, {% endif %}
+    {% endfor %}
+                ],
+    pointBorderColor: '#01C696',
+        pointBackgroundColor: '#ffffff',
+            fontColor: '#ffffff',
+            }]
+        },
+
+    // Configuration options go here
+    options: {
+        maintainAspectRatio: false,
+            responsive: true,
+                tooltips: {
+            mode: 'index',
+                intersect: false,
+            },
+        hover: {
+            mode: 'nearest',
+                intersect: true
+        },
+        scales: {
+            xAxes: [{
+                type: 'time',
+                time: {
+                    format: 'DD/MM/YYYY',
+                    unit: 'day'
+                },
+                display: true,
+                scaleLabel: {
+                    display: true,
+                    labelString: 'Last 30 days',
+                    padding: 1
+                }
+            }]
+        }
+    }
+    });
+</script>
